substituição do reload por useEffect corretamente, agora ao fazer pedido é levado para a tela home com o pedido do acompanhamente a ser exibido
