+++ src/pages/ProfileEditAdress/ProfileEditAdress.js
@@ -35,10 +35,8 @@ const ProfileEditAdress = () => {
     })
     .then((res) => {
       localStorage.setItem('token', res.data.token)
-
       alert('Seu endereço foi alterado com sucesso!');
       navigate('/profile');
-      reload()
     })
     .catch((err) => {
       alert(err.response.data);
