+++ src/pages/Register/RegisterForm.js
@@ -41,10 +41,10 @@ const RegisterForm = () => {
       .post(`${BASE_URL}signup`, body)
       .then(res => {
         localStorage.setItem('token', res.data.token)
-        console.log('Deu mt bom')
+        alert("Preencha seu endereço")
         navigate('/profile/editAdress')
       })
-      .catch((err) => {alert(err.response.data.message)}
+      .catch((err) => {console.log(err.response.data.message)}
       )
   }
